+++ src/views-components/project-copy-project-dialog/project-copy-project-dialog.tsx
@@ -7,14 +7,12 @@ import { withDialog } from "~/store/dialog/with-dialog";
 import { reduxForm, startSubmit, stopSubmit } from 'redux-form';
 import { PROJECT_COPY_DIALOG } from '~/store/project-copy-project-dialog/project-copy-project-dialog';
 import { ProjectCopyFormDialog } from "~/views-components/project-copy-dialog/project-copy-dialog";
+import { copyProject } from '../../store/project-copy-project-dialog/project-copy-project-dialog';
 
 export const ProjectCopyDialog = compose(
     withDialog(PROJECT_COPY_DIALOG),
     reduxForm({
         form: PROJECT_COPY_DIALOG,
-        onSubmit: (data, dispatch) => {
-            dispatch(startSubmit(PROJECT_COPY_DIALOG));
-            setTimeout(() => dispatch(stopSubmit(PROJECT_COPY_DIALOG, { name: 'Invalid path' })), 2000);
-        }
+        onSubmit: (data, dispatch) => dispatch(copyProject(data))
     })
 )(ProjectCopyFormDialog);
